requirements.txt
absl-py==1.4.0
apache-tvm==0.10.0
appdirs==1.4.4
arrow==1.2.3
astunparse==1.6.3
attrs==22.2.0
black==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
coloredlogs==15.0.1
decorator==5.1.1
dill==0.3.6
dm-tree==0.1.8
docopt==0.6.2
filelock==3.9.0
flatbuffers==23.3.3
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
huggingface-hub==0.12.1
humanfriendly==10.0
hummingbird-ml==0.4.8
idna==3.4
importlib-metadata==6.0.0
inform==1.27.0
joblib==1.2.0
keras==2.11.0
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
mpmath==1.2.1
mypy-extensions==1.0.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
onnx==1.13.1
onnx-tf==1.10.0
onnxconverter-common==1.13.0
onnxruntime==1.14.1
opt-einsum==3.3.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.0
protobuf==3.20.3
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
PyYAML==6.0
quantiphy==2.19
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
sacremoses==0.0.53
scikit-learn==1.2.1
scipy==1.10.1
sentencepiece==0.1.97
six==1.16.0
sympy==1.11.1
synr==0.6.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-addons==0.19.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-probability==0.19.0
termcolor==2.2.0
threadpoolctl==3.1.0
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.1
tornado==6.2
tqdm==4.65.0
transformers==4.26.1
typeguard==2.13.3
typing-extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
